+++ Return.h
@@ -14,7 +14,7 @@ public:
 
   ~Return();
 
-  void add(string dr, double amount, string returner, string bc, string dB);
+  void add(string dr, int amount, string returner, string bc, string dB);
   void updateReturn(int amount, string newDateReturned);
   int getReturn();
   string getDatePurchased();
