+++ dialog.cpp
@@ -232,7 +232,7 @@ void Dialog::on_Buy_clicked()
                     getAmount(i);
             string price = multiBuyController->getPrice(i);
             inventory->getVegetableByIndex(vegeIndex)->buyVege(
-                        stod(amount), company, date->text().toStdString(),
+                        stoi(amount), company, date->text().toStdString(),
                         price);
         }
 
@@ -690,7 +690,7 @@ static int textWidth = 380;
 static int pageHeight = 1000;
 
 void Dialog::printT(QPrinter* printer){
-    double amount;
+    int amount;
     time_t t = time(nullptr);
     struct tm * now = localtime(&t);
     char buffer[128];
@@ -816,7 +816,7 @@ void Dialog::printT(QPrinter* printer){
 }
 
 void Dialog::printI(QPrinter* printer){
-    double amount;
+    int amount;
     time_t t = time(nullptr);
     struct tm * now = localtime(&t);
     char buffer[128];
@@ -948,7 +948,7 @@ void Dialog::printH(QPrinter * printer){
 
     //qdialog what day do you want to print? have a drop down box with all the available days
     string dayComp;
-    double amount;
+    int amount;
     time_t t = time(nullptr);
     struct tm * now = localtime(&t);
     char buffer[128];
@@ -1300,7 +1300,7 @@ void Dialog::dumpVege(){
         QObject::connect(&checkBox,SIGNAL(toggled(bool)), &dialog, SLOT(reject()));
 
         int result = dialog.exec();
-        double amount = lineEdit->text().toInt();
+        int amount = lineEdit->text().toInt();
         int selection = remainingDrop->currentIndex();
 
         time_t t = time(nullptr);
@@ -1611,7 +1611,7 @@ void Dialog::on_Return_clicked()
         QObject::connect(&checkBox,SIGNAL(toggled(bool)), &dialog, SLOT(reject()));
 
         int result = dialog.exec();
-        double amount = lineEdit->text().toInt();
+        int amount = lineEdit->text().toInt();
         string dateBought = lineEdit2->text().toUtf8().constData();
         string returner = customerDrop->currentText().toUtf8().constData();
         string company = companyDrop->currentText().toUtf8().constData();
@@ -1661,7 +1661,7 @@ void Dialog::on_Return_clicked()
                 QObject::connect(&buttonBox, SIGNAL(rejected()), &dialog, SLOT(reject()));
 
                 int result = dialog.exec();
-                double amount = lineEdit->text().toInt();
+                int amount = lineEdit->text().toInt();
                 int selection = remainingDrop->currentIndex();
 
                 time_t t = time(nullptr);
@@ -1745,7 +1745,7 @@ void Dialog::on_dumpButton_clicked()
         QObject::connect(&checkBox,SIGNAL(toggled(bool)), &dialog, SLOT(reject()));
 
         int result = dialog.exec();
-        double amount = lineEdit->text().toInt();
+        int amount = lineEdit->text().toInt();
         int selection = remainingDrop->currentIndex();
 
         time_t t = time(nullptr);
@@ -1922,7 +1922,7 @@ void Dialog:: undoHistory(){
                                 mTranslator->translate("你要打勾所有的選項才能undo!").c_str());
             messageBox.setFixedSize(500,200);
         }else{
-            double amount = 0;
+            int amount = 0;
             string dP, company, customer, dS;
             amount = currentVege->getHistoryObject(ui->historyList->currentRow())->getDifference();
             dP =  currentVege->getHistoryObject(ui->historyList->currentRow())->getDatePurchased();
@@ -2051,7 +2051,7 @@ void Dialog::on_clearHistoryButton_clicked()
         QObject::connect(&checkBox,SIGNAL(toggled(bool)), &dialog, SLOT(reject()));
 
         int result = dialog.exec();
-        double amount = lineEdit->text().toInt();
+        int amount = lineEdit->text().toInt();
 
 
         // Show the dialog as modal
